Refactor LogoDisplay - Add Logo only during Feature
